+++ WheelWizard/Features/WiiManagement/GameLicense/Domain/Statistics/WinLoss.cs
@@ -1,0 +1,7 @@
+namespace WheelWizard.WiiManagement.GameLicense.Domain.Statistics;
+
+public class WinLoss
+{
+    public uint Wins { get; set; }
+    public uint Losses { get; set; }
+}
