+++ WheelWizard/Features/WiiManagement/MiiManagement/MiiRepositoryService.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO.Abstractions;
+using WheelWizard.Helpers;
 using WheelWizard.Services;
-using WheelWizard.Services.WiiManagement.SaveData;
 
 namespace WheelWizard.WiiManagement.MiiManagement;
 
@@ -132,7 +132,7 @@ public class MiiRepositoryServiceService(IFileSystem fileSystem) : IMiiRepositor
             if (block.Length != MiiLength)
                 continue;
 
-            var thisId = BigEndianBinaryReader.BufferToUint32(block, 0x18);
+            var thisId = BigEndianBinaryHelper.BufferToUint32(block, 0x18);
             if (thisId == clientId)
                 return block;
         }
@@ -198,7 +198,7 @@ public class MiiRepositoryServiceService(IFileSystem fileSystem) : IMiiRepositor
             if (block.Length != MiiLength)
                 continue;
 
-            var thisId = BigEndianBinaryReader.BufferToUint32(block, 0x18);
+            var thisId = BigEndianBinaryHelper.BufferToUint32(block, 0x18);
             if (thisId != clientId)
                 continue;
