+++ WheelWizard/Features/WiiManagement/MiiManagement/MiiSerializer.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text;
-using WheelWizard.Services.WiiManagement.SaveData;
+using WheelWizard.Helpers;
 using WheelWizard.WiiManagement.MiiManagement.Domain.Mii;
 
 namespace WheelWizard.WiiManagement.MiiManagement;
@@ -39,7 +39,7 @@ public static class MiiSerializer
         data[0x17] = mii.Weight.Value;
 
         // Mii ID (0x18 - 0x1B)
-        BigEndianBinaryReader.WriteUInt32BigEndian(data, 0x18, mii.MiiId);
+        BigEndianBinaryHelper.WriteUInt32BigEndian(data, 0x18, mii.MiiId);
 
         // System ID (0x1C - 0x1F)
         data[0x1C] = mii.SystemId0;
@@ -186,7 +186,7 @@ public static class MiiSerializer
         mii.Weight = weight.Value;
 
         // Mii ID (0x18 - 0x1B)
-        mii.MiiId = BigEndianBinaryReader.BufferToUint32(data, 0x18);
+        mii.MiiId = BigEndianBinaryHelper.BufferToUint32(data, 0x18);
 
         // System ID (0x1C - 0x1F)
         mii.SystemId0 = data[0x1C];
