+++ WheelWizard/Utilities/Generators/FriendCodeGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
-using WheelWizard.Services.WiiManagement.SaveData;
+using WheelWizard.Helpers;
 
 namespace WheelWizard.Utilities.Generators;
 
@@ -8,7 +8,7 @@ public class FriendCodeGenerator
 {
     public static string GetFriendCode(byte[] data, int offset)
     {
-        var pid = BigEndianBinaryReader.BufferToUint32(data, offset);
+        var pid = BigEndianBinaryHelper.BufferToUint32(data, offset);
         if (pid == 0)
             return string.Empty;
